Feat(cmd/list): Add filtering for RFC IDs.
This means that now RFC file names will be filtered for containing a numeric ID
of 3 or more digits, as well as a text-document file extension, meaning that
`001.md` will match, `some-other-format-192.rst` will match, but not
`README.md`, for example. As with everything, this is a heuristic, so it will
fail at least some of the time.