feat(cmd/create/next_rfc_number: Clarify folding logic
While the old version also worked perfectly fine, it more or less worked on
accident and also looked like a bug, despite working perfectly fine. Moving the
increment to happen *after* the fold makes the logic flow clearer.

This prompted by me stumbling over this while reading and needing to
triple-check if this was or was not a bug, and coming to the conclusion that it
was fine, just hard to read.